requirements.txt
Flask==2.1.3
imutils==0.5.4
Keras==2.4.0
opencv-python==4.4.0.46
Pillow==9.2.0
tensorflow==2.3.1
scikit-image==0.16.2
